% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReadSpatial.R
\name{dbReadSpatial}
\alias{dbReadSpatial}
\title{Read Spatial Data from SQL Database}
\usage{
dbReadSpatial(
  conn,
  schema.name = NULL,
  table.name,
  geom.col = "geom_wkt",
  default.srid = 4326L,
  dialect
)
}
\arguments{
\item{conn}{A valid database connection object.}

\item{schema.name}{Schema name. Defaults based on dialect.}

\item{table.name}{Table name.}

\item{geom.col}{Name of the WKT geometry column. Default is "geom_wkt".}

\item{default.srid}{SRID to use if not embedded in WKT. Default is `4326`.}

\item{dialect}{Character string specifying database dialect: "mssql" or "postgres".}
}
\value{
An `sf` object.
}
\description{
Reads a table containing WKT geometries and converts back to an sf object.
}
\examples{
\dontrun{
nc.sf <- dbReadSpatial(
  conn = conn,
  table.name = "nc_counties",
  dialect = "postgres"
)
}
}
