% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifyTableMs.R
\name{qualifyTableMs}
\alias{qualifyTableMs}
\title{Qualify schema + table name safely}
\usage{
qualifyTableMs(schema, table)
}
\arguments{
\item{schema}{schema name}

\item{table}{table name}
}
\value{
[schema].[table]
}
\description{
Qualify schema + table name safely
}
