% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualifyTablePg.R
\name{qualifyTablePg}
\alias{qualifyTablePg}
\title{Qualify schema + table name safely for PostgreSQL}
\usage{
qualifyTablePg(schema, table)
}
\arguments{
\item{schema}{schema name}

\item{table}{table name}
}
\value{
"schema"."table"
}
\description{
Qualify schema + table name safely for PostgreSQL
}
