% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoteIdentMs.R
\name{quoteIdentMs}
\alias{quoteIdentMs}
\title{Quote SQL Server identifier safely using []
Escapes ']' as ']]'}
\usage{
quoteIdentMs(x)
}
\arguments{
\item{x}{character vector}
}
\value{
quoted identifiers
}
\description{
Quote SQL Server identifier safely using []
Escapes ']' as ']]'
}
