% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoteIdentPg.R
\name{quoteIdentPg}
\alias{quoteIdentPg}
\title{Quote PostgreSQL identifier safely using double quotes
Escapes '"' as '""'}
\usage{
quoteIdentPg(x)
}
\arguments{
\item{x}{character vector}
}
\value{
quoted identifiers
}
\description{
Quote PostgreSQL identifier safely using double quotes
Escapes '"' as '""'
}
